Cambiando el tipo de dato idLiquidacion a long, se crea index.html

+++ src/main/java/cl/awakelab/leandrovillalba/sprint6/service/ILiquidacionService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface ILiquidacionService {
     Liquidacion crearLiquidacion(Liquidacion liquidacion);
     List<Liquidacion> listarLiquidaciones();
-    Liquidacion buscarLiquidacionPorId(int idLiquidacion);
-    Liquidacion actualizarLiquidacion(Liquidacion liquidacion, int idLiquidacion);
+    Liquidacion buscarLiquidacionPorId(long idLiquidacion);
+    Liquidacion actualizarLiquidacion(Liquidacion liquidacion, long idLiquidacion);
     Liquidacion actualizarLiquidacion2(Liquidacion liquidacionActualizar);
     void eliminarLiquidacion(Liquidacion liquidacion);
-    void eliminarLiquidacion2(int idLiquidacion);
+    void eliminarLiquidacion2(long idLiquidacion);
 }
